requirements.txt
PyMuPDF==1.23.21
google-generativeai==0.3.2
python-dotenv==1.0.1 